Trait pallet_mt::pallet::Config

pub trait Config<I: 'static = ()>: Config {

    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type TreeId: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen;
    type LeafIndex: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen;
    type RootIndex: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen;
    type Element: ElementTrait + MaxEncodedLen;
    type DefaultZeroElement: Get<Self::Element>;
    type MaxTreeDepth: Get<u8>;
    type Hasher: HasherModule;
    type Currency: ReservableCurrency<Self::AccountId>;
    type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type TreeDeposit: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>;
    type DataDepositBase: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>;
    type DataDepositPerByte: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>;
    type Two: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>;
    type RootHistorySize: Get<Self::RootIndex>;
    type StringLimit: Get<u32>;
    type MaxEdges: Get<u32> + TypeInfo;
    type MaxDefaultHashes: Get<u32> + TypeInfo;
    type WeightInfo: WeightInfo;

}

The module configuration trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type TreeId: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen

The overarching tree ID type

type LeafIndex: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen

The overarching leaf index type

type RootIndex: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen

The overarching leaf index type

type Element: ElementTrait + MaxEncodedLen

the leaf type

type DefaultZeroElement: Get<Self::Element>

the default zero element

type MaxTreeDepth: Get<u8>

The max depth of trees

type Hasher: HasherModule

The hasher instance trait

type Currency: ReservableCurrency<Self::AccountId>

The currency mechanism.

type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin which may forcibly modify the tree

type TreeDeposit: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>

The basic amount of funds that must be reserved for an tree.

type DataDepositBase: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>

The basic amount of funds that must be reserved when adding metadata to your parameters.

type DataDepositPerByte: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>

The additional funds that must be reserved for the number of bytes you store in your parameter metadata.

type Two: Get<<<Self as Config<I>>::Currency as Currency<<Self as SystemConfig>::AccountId>>::Balance>

The value of two in this form

type RootHistorySize: Get<Self::RootIndex>

History size of roots for each tree

type StringLimit: Get<u32>

The maximum length of a name or symbol stored on-chain.

type MaxEdges: Get<u32> + TypeInfo

The maximum count of edges nodes to be stored for a tree detail

type MaxDefaultHashes: Get<u32> + TypeInfo

The maximum count of default hashes to store

type WeightInfo: WeightInfo

WeightInfo for pallet